benchmark/pyproject.toml
[build-system]
requires = ["uv_build>=0.9.15,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
  "pdbpp>=0.11.7"
]
hf = [
  "transformers>=4.57.3"
]

[project]
name = "benchmark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
  {name = "Louis Maddox", email = "louismmx@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
  "polars-fastembed"
]

[project.scripts]
benchmark = "benchmark:main"
count-pep-tokens = "benchmark.tokens:count_pep_tokens"

[tool.uv]
default-groups = ["dev", "hf"]

[tool.uv.sources]
polars-fastembed = {path = "../polars-fastembed", editable = true}
